set_signal_value_long() bugfixes and improvements (#2316)
* Issue #1202: Change gpi_set_signal_value_long(long) to gpi_set_signal_value_int(int32_t)

- vpi_put_value() is limited to 32-bit, so use int32_t instead of long.
- Fix 32-bit platform compilation warning: comparison of integer expressions of different signedness.

(cherry picked from commit 99079fa70b5f7b41bd68f52dd1ec9fc0e4e3b4b2)

* Issue #1202: set IntegerObject using set_signal_value_int()

(cherry picked from commit 9080c0d3c9fb575cfffc6aab573cca513a5f4695)

* Issue #1202: Improve setting IntegerObject values

- Check if value for IntegerObject is in 32-bit range
- Raise OverflowError for out of range values
- Update integer tests for the supported range and out of range

(cherry picked from commit b15be8a66bc0490e30df00183739f634d14ef895)

* Issue #1202: Replace set_signal_val_long() with set_signal_val_int()

(cherry picked from commit 2ac766f4f59b4f14ea91fb5c130bce29d31228c2)

* Issue #913: Fix issue setting ints to signals wider than 32-bit

- Fix handling of negative numbers for signal widths wider than 32-bit
- Add value range check
- Add testcases
- Add documentation about signed and unsigned values and supported range
- Add newsfragment

(cherry picked from commit 8548afccf66158db7a1402d957ba4e96e0166012)

* Issue #268: use full 32-bit range to set signals

- Performance improvement for setting 32-bit wide signals.
- For signals up to 32-bit wide, set_signal_val_int() is used to set the full
  signed min negative to unsigned max positive data range from a python integer value.
- Add max supported signal width check and signal width dependent overflow check in set_signal_val_int().

(cherry picked from commit 6e56a5e3da111412c1eab9e717bd703944e2e6a4)

* Address review comments:

- Consolidate integer and int tests and use assert_raises to check OverflowError
- Perform value range check in handle.py rather than simulatormodule.cpp
- Fix riviera regression test failures
- Add newsfragment

+++ cocotb/share/lib/fli/FliImpl.h
@@ -240,7 +240,7 @@ public:
     double get_signal_value_real() override;
     long get_signal_value_long() override;
 
-    int set_signal_value(long value, gpi_set_action_t action) override;
+    int set_signal_value(int32_t value, gpi_set_action_t action) override;
     int set_signal_value(double value, gpi_set_action_t action) override;
     int set_signal_value_str(std::string &value, gpi_set_action_t action) override;
     int set_signal_value_binstr(std::string &value, gpi_set_action_t action) override;
@@ -277,7 +277,7 @@ public:
     long get_signal_value_long() override;
 
     using FliValueObjHdl::set_signal_value;
-    int set_signal_value(long value, gpi_set_action_t action) override;
+    int set_signal_value(int32_t value, gpi_set_action_t action) override;
 
     int initialise(std::string &name, std::string &fq_name) override;
 
@@ -316,7 +316,7 @@ public:
     const char* get_signal_value_binstr() override;
 
     using FliValueObjHdl::set_signal_value;
-    int set_signal_value(long value, gpi_set_action_t action) override;
+    int set_signal_value(int32_t value, gpi_set_action_t action) override;
     int set_signal_value_binstr(std::string &value, gpi_set_action_t action) override;
 
     int initialise(std::string &name, std::string &fq_name) override;
@@ -347,7 +347,7 @@ public:
     long get_signal_value_long() override;
 
     using FliValueObjHdl::set_signal_value;
-    int set_signal_value(long value, gpi_set_action_t action) override;
+    int set_signal_value(int32_t value, gpi_set_action_t action) override;
 
     int initialise(std::string &name, std::string &fq_name) override;
 };
