+++ cocotb/share/lib/gpi/gpi_priv.h
@@ -154,7 +154,7 @@ public:
 
     int m_length = 0;
 
-    virtual int set_signal_value(const long value, gpi_set_action_t action) = 0;
+    virtual int set_signal_value(const int32_t value, gpi_set_action_t action) = 0;
     virtual int set_signal_value(const double value, gpi_set_action_t action) = 0;
     virtual int set_signal_value_str(std::string &value, gpi_set_action_t action) = 0;
     virtual int set_signal_value_binstr(std::string &value, gpi_set_action_t action) = 0;
