+++ cocotb/share/lib/vpi/VpiImpl.h
@@ -179,7 +179,7 @@ public:
     double get_signal_value_real() override;
     long get_signal_value_long() override;
 
-    int set_signal_value(const long value, gpi_set_action_t action) override;
+    int set_signal_value(const int32_t value, gpi_set_action_t action) override;
     int set_signal_value(const double value, gpi_set_action_t action) override;
     int set_signal_value_binstr(std::string &value, gpi_set_action_t action) override;
     int set_signal_value_str(std::string &value, gpi_set_action_t action) override;
